+++ depynd/markov_networks/__init__.py
@@ -1,5 +1,6 @@
 from depynd.markov_networks.skeptic_ import skeptic
 from depynd.markov_networks.gsmn_ import gsmn
+from depynd.markov_networks.iamb_ import iamb
 from depynd.markov_networks.select_ import stars, select
 
-__all__ = ['skeptic', 'gsmn', 'stars', 'select']
+__all__ = ['skeptic', 'gsmn', 'iamb', 'stars', 'select']
